+++ src/helpers.h
@@ -149,4 +149,15 @@ Eigen::VectorXd gr_nloglr_qpoi_log(const Eigen::Ref<const Eigen::VectorXd> &l,
                                    const Eigen::Ref<const Eigen::VectorXd> &par,
                                    const Eigen::Ref<const Eigen::ArrayXd> &w,
                                    const bool weighted);
+
+Eigen::MatrixXd g_qpoi_identity(const Eigen::Ref<const Eigen::MatrixXd> &x,
+                                const Eigen::Ref<const Eigen::VectorXd> &par);
+
+Eigen::VectorXd gr_nloglr_qpoi_identity(
+    const Eigen::Ref<const Eigen::VectorXd> &l,
+    const Eigen::Ref<const Eigen::MatrixXd> &g,
+    const Eigen::Ref<const Eigen::MatrixXd> &x,
+    const Eigen::Ref<const Eigen::VectorXd> &par,
+    const Eigen::Ref<const Eigen::ArrayXd> &w,
+    const bool weighted);
 #endif