+++ tests/testthat/test-el_glm.R
@@ -23,7 +23,7 @@ test_that("Invalid `family`.", {
     data = airquality
   ))
   expect_error(el_glm(carb ~ .,
-    family = quasipoisson("identity"),
+    family = quasipoisson("sqrt"),
     data = mtcars
   ))
 })